Separated the tests and got coverage for the tests... these are the first two tests...
